vanaadhikar-drishti/backend/pyproject.toml
[project]
name = "fra-legal-processor"
version = "0.1.0"
description = "Production-grade multilingual document ingestion service for Vanaadhikar Drishti"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.5",
  "loguru>=0.7.2",
  "pandas>=2.2.3",
  "pdfplumber>=0.11.5",
  "pillow>=11.0.0",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.7.0",
  "pymysql>=1.1.0",
  "python-multipart>=0.0.9",
  "python-slugify>=8.0.4",
  "python-jose[cryptography]>=3.3.0",
  "sqlalchemy>=2.0.36",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.32.1",
  "python-docx>=1.1.2",
  "pytesseract>=0.3.13",
  "opencv-python-headless>=4.10.0.84",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
